Support for cardnumber only
We can now set that we only want the cardnumber using optional protocols.
We can also be informed when the cardnumber is valid, without having to
show the meta views, also using optional protocols.

So this shouldn't affect existing users, only give more options to the
whomever needs it. (Me for example, I need this :))