slide back to card number only when it's invalid
Relates to https://github.com/stripe/PaymentKit/issues/39

This is a behaviour change. Previously, when a user deleted the last (leftmost) character of the expiration date and the focus moved to the card number, there was no intuitive way to return to the date field, since it slid away to the right. The user could delete the last (rightmost) character of the card number and re-type it, but that's hardly intuitive.

The new behaviour is that when the focus moves to card number field the other fields stay in view and can easily be tapped. Only when the last (rightmost) character of the card number field is deleted does the animation left happen. If the user then re-types the last character, the animation right happens and focus moves to the expiration date field.