char after end of valid field rolls over to next
When caret is after last char of valid CC number field and user enters more text it's now inserted at the beginning of the expiration date field (as long as it's valid there).
Same goes for the end of the expiration date field and the beginning of the CVC field.

The next field also becomes nextResponder.